Code/requirements.txt
torch
transformers
sympy
scipy
joblib
scikit-learn
deep_translator
nltk
python-dotenv
